+++ internal/authorization_provider.go
@@ -1,0 +1,9 @@
+package internal
+
+const (
+	authorizationHeader = "Authorization"
+)
+
+type AuthorizationProvider interface {
+	AuthorizeRequest(request RequestInformation) error
+}