+++ internal/credential.go
@@ -1,0 +1,5 @@
+package internal
+
+type Credential interface {
+	GetAuthentication() (string, error)
+}