+++ servicenow_client.go
@@ -10,12 +10,15 @@ import (
 	"strings"
 
 	"github.com/michaeldcanady/servicenow-sdk-go/core"
+	"github.com/michaeldcanady/servicenow-sdk-go/internal"
 )
 
 type ServiceNowClient struct {
-	Credential core.Credential
-	BaseUrl    string //nolint:stylecheck
-	Session    http.Client
+	// Deprecated: deprecated since v{unreleased}.
+	Credential   core.Credential
+	authProvider *internal.BaseAuthorizationProvider
+	BaseUrl      string //nolint:stylecheck
+	Session      http.Client
 }
 
 // Now returns a NowRequestBuilder associated with the Client.
@@ -24,6 +27,7 @@ func (c *ServiceNowClient) Now() *NowRequestBuilder {
 	return NewNowRequestBuilder(c.BaseUrl+"/now", c)
 }
 
+// Deprecated: deprecated since v{unreleased}. Please use `NewServiceNowClient2` instead.
 // NewServiceNowClient creates a new instance of the ServiceNow client.
 // It accepts a UsernamePasswordCredential and an instance URL.
 // If the instance URL does not end with ".service-now.com/api", it appends the suffix.
@@ -37,11 +41,40 @@ func NewServiceNowClient(credential core.Credential, instance string) *ServiceNo
 		instance = "https://" + instance
 	}
 
+	authProvider, _ := internal.NewBaseAuthorizationProvider(credential)
+
 	return &ServiceNowClient{
-		Credential: credential,
-		BaseUrl:    instance,
-		Session:    http.Client{},
+		Credential:   credential,
+		authProvider: authProvider,
+		BaseUrl:      instance,
+		Session:      http.Client{},
 	}
+}
+
+// NewServiceNowClient2 creates a new instance of the ServiceNow client.
+// It accepts a UsernamePasswordCredential and an instance URL.
+// If the instance URL does not end with ".service-now.com/api", it appends the suffix.
+// It returns a pointer to the Client.
+func NewServiceNowClient2(credential core.Credential, instance string) (*ServiceNowClient, error) {
+	if !strings.HasSuffix(instance, ".service-now.com/api") {
+		instance += ".service-now.com/api"
+	}
+
+	if !strings.HasPrefix(instance, "https://") {
+		instance = "https://" + instance
+	}
+
+	authProvider, err := internal.NewBaseAuthorizationProvider(credential)
+	if err != nil {
+		return nil, err
+	}
+
+	return &ServiceNowClient{
+		Credential:   credential,
+		authProvider: authProvider,
+		BaseUrl:      instance,
+		Session:      http.Client{},
+	}, nil
 }
 
 func (c *ServiceNowClient) unmarshallError(response *http.Response) error {
@@ -96,17 +129,16 @@ func (c *ServiceNowClient) toRequestWithContext(ctx context.Context, requestInfo
 		return nil, ErrNilContext
 	}
 
+	err := c.authProvider.AuthorizeRequest(requestInfo)
+	if err != nil {
+		return nil, err
+	}
+
 	request, err := requestInfo.ToRequestWithContext(ctx)
 	if err != nil {
 		return nil, err
 	}
 
-	authHeader, err := c.Credential.GetAuthentication()
-	if err != nil {
-		return nil, err
-	}
-
-	request.Header.Add("Authorization", authHeader)
 	request.Header.Add("Content-Type", "application/json")
 	request.Header.Add("Accept", "application/json")
 	return request, nil