+++ servicenow_client_test.go
@@ -6,15 +6,34 @@ import (
 	"io"
 	"net/http"
 	"net/url"
+	"reflect"
 	"strings"
 	"testing"
 
 	"github.com/michaeldcanady/servicenow-sdk-go/core"
 	"github.com/michaeldcanady/servicenow-sdk-go/credentials"
+	"github.com/michaeldcanady/servicenow-sdk-go/internal"
+	"github.com/mozillazg/go-httpheader"
 	"github.com/stretchr/testify/assert"
 )
 
+type test[T any] struct {
+	Title string
+	// Setup to make needed modifications for a specific test
+	Setup func()
+	// Cleanup to undo changes do to reusable items
+	Cleanup     func()
+	Input       interface{}
+	Expected    T
+	expectedErr error
+}
+
+var (
+	sharedUsernameAndPasswordCred = credentials.NewUsernamePasswordCredential("username", "password")
+)
+
 type MockRequestInformation struct {
+	Headers http.Header
 }
 
 func (rI *MockRequestInformation) AddRequestOptions(options []core.RequestOption) {
@@ -44,10 +63,42 @@ func (rI *MockRequestInformation) ToRequest() (*http.Request, error) {
 }
 
 func (rI *MockRequestInformation) ToRequestWithContext(ctx context.Context) (*http.Request, error) {
-	return http.NewRequestWithContext(ctx, "GET", "https://www.example.com", nil)
+	request, err := http.NewRequestWithContext(ctx, "GET", "https://www.example.com", nil)
+	if err != nil {
+		return nil, err
+	}
+	request.Header = rI.Headers
+
+	return request, nil
 }
 
 func (rI *MockRequestInformation) AddHeaders(rawHeaders interface{}) error {
+	var headers http.Header
+	var err error
+
+	val := reflect.ValueOf(rawHeaders)
+
+	if val.Kind() == reflect.Struct {
+		// use the httpheader.Encode function from the httpheader package
+		// to encode the pointer value into an http.Header map
+		headers, err = httpheader.Encode(rawHeaders)
+		if err != nil {
+			return err
+		}
+	} else if val.Type() == reflect.TypeOf(http.Header{}) {
+		// if the value is already an http.Header map, just assign it to headers
+		headers = rawHeaders.(http.Header)
+	} else {
+		// otherwise, return an error
+		return core.ErrInvalidHeaderType
+	}
+
+	// iterate over the headers map and add each key-value pair to rI.Headers
+	for key, values := range headers {
+		for _, value := range values {
+			rI.Headers.Add(key, value)
+		}
+	}
 	return nil
 }
 
@@ -57,6 +108,43 @@ func TestNewClient(t *testing.T) {
 	client := NewServiceNowClient(cred, "instance")
 
 	assert.NotNil(t, client)
+}
+
+func TestNewClient2(t *testing.T) {
+	authProvider, _ := internal.NewBaseAuthorizationProvider(sharedUsernameAndPasswordCred)
+
+	tests := []test[*ServiceNowClient]{
+		{
+			Title: "Valid",
+			Input: []interface{}{"instance", sharedUsernameAndPasswordCred},
+			Expected: &ServiceNowClient{
+				Credential:   sharedUsernameAndPasswordCred,
+				authProvider: authProvider,
+				BaseUrl:      "https://instance.service-now.com/api",
+				Session:      http.Client{},
+			},
+			expectedErr: nil,
+		},
+		{
+			Title:       "Nil Credential",
+			Input:       []interface{}{"instance", (*credentials.UsernamePasswordCredential)(nil)},
+			Expected:    nil,
+			expectedErr: internal.ErrNilCredential,
+		},
+	}
+
+	for _, test := range tests {
+		t.Run(test.Title, func(t *testing.T) {
+			values := test.Input.([]interface{})
+
+			instance := values[0].(string)
+			cred := values[1].(core.Credential)
+
+			client, err := NewServiceNowClient2(cred, instance)
+			assert.Equal(t, test.Expected, client)
+			assert.Equal(t, test.expectedErr, err)
+		})
+	}
 }
 
 func TestClientURL(t *testing.T) {
@@ -80,7 +168,9 @@ func TestClientNow(t *testing.T) {
 }
 
 func TestClientToRequest(t *testing.T) {
-	requestInfo := &MockRequestInformation{}
+	requestInfo := &MockRequestInformation{
+		Headers: http.Header{},
+	}
 
 	cred := credentials.NewUsernamePasswordCredential("username", "password")
 
@@ -141,7 +231,9 @@ func TestClientUnmarshallError(t *testing.T) {
 }
 
 func TestClientToRequestWithContext(t *testing.T) {
-	requestInfo := &MockRequestInformation{}
+	requestInfo := &MockRequestInformation{
+		Headers: http.Header{},
+	}
 
 	cred := credentials.NewUsernamePasswordCredential("username", "password")