+++ changelog.md
@@ -15,9 +15,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Deprecated
 
+- `servicenowsdkgo.ServiceNowClient.Credential`
+
 ### Removed
 
 ### Fixed
+
+- Improperly embedded `core.PageIterator2[T]` in `tableapi.TablePageIterator` ([Issue #105](https://github.com/michaeldcanady/servicenow-sdk-go/issues/105))
 
 ## [1.5.0] - (02/03/2024)